Text Edit Update
Everybody thank Jorrit on the EaWX Team for making this cool tool. If you wanna edit text for the mod, add a new line in MasterTextFile_English.txt and run the alphabetize and build bat file. This will sort and export the text file into the dat files, no more dumb dat editor!